moving gptbehaviortable to gptbehavior app
