+++ package.json
@@ -29,6 +29,7 @@
     "vue": "^3.2.13",
     "vue-moment": "^4.1.0",
     "vue-router": "^4.2.2",
+    "vue3-easy-data-table": "^1.5.47",
     "vue3-table-lite": "^1.3.7",
     "vuex": "^4.0.2"
   },
